Deleten korjauksia

+++ app/controllers/material_controller.php
@@ -72,9 +72,15 @@ class MaterialController extends BaseController {
     public static function destroy($id) {
         self::check_logged_in();
         $material = new Material(array('id' => $id));
+        $ok_to_delete = $material->can_be_deleted();
+        
+        if ($ok_to_delete){
         $material->destroy();
 
         Redirect::to('/material', array('message' => 'Materiaali on poistettu onnistuneesti!'));
+        } else {
+            Redirect::to('/material/' . $id, array('errors' => 'Materiaali on liitetty tentteihin, joten sitä ei voi poistaa.'));
+        }
     }
 
     public static function create() {
